sof-event-extractor/backend/requirements.txt
flask==3.1.2
flask-cors==5.0.0
azure-storage-blob==12.26.0
requests==2.32.5
# OCR and document processing
pytesseract==0.3.10
pdf2image==1.17.0
python-docx==1.1.2
Pillow==10.4.0
# NLP processing (install separately if needed)
# spacy==3.8.2
# Optional: pyodbc for SQL Server (install system drivers separately)
# pyodbc==5.2.0
